% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculator.R
\name{get_scores}
\alias{get_scores}
\title{Calculate IDEFICS Scores for Children}
\usage{
get_scores(
  variable = "waist",
  sex = c("f", "m"),
  age = 6:5,
  height = NULL,
  values = c(20, 21),
  return_values = c("percentile", "z.score")
)
}
\arguments{
\item{variable}{Character. The variable to assess. Must be one of the supported variables ("waist", "bmi", "hdl", "sbp", "dbp", "trg", "homa", "glu", "height" or "insu").}

\item{sex}{Character vector. Same length as `age`, `height`, and `values`. Accepts "f" for female or "m" for male.}

\item{age}{Numeric vector. Ages of the children in years.}

\item{height}{Numeric vector or NULL. Required for height-dependent models (`sbp` and `dbp`). Defaults to NULL.}

\item{values}{Numeric vector. Observed values of the variable to score.}

\item{return_values}{Character vector. Specifies which outputs to return. Options include "percentile", "z.score".}
}
\value{
A named list containing the requested scores. Each element is a numeric vector of the same length as `values`.
}
\description{
Computes age-, sex- and (for `sbp` and `dbp`) height-specific percentiles or z-scores for anthropometric and metabolic variables
using IDEFICS study reference data.
}
\examples{

get_scores(
  variable="dbp",
  sex=c("f","m"),
  age=c(5,5),
  height=c(120,110),
  values=c(70,60)
)

}
