% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculator.R
\name{p_hurdle}
\alias{p_hurdle}
\title{Function to calculate percentiles from the hurdle model}
\usage{
p_hurdle(y, p1, mu, sigma, nu, tau)
}
\arguments{
\item{y}{Numeric vector. Observed values of the variable to score.}

\item{p1}{cutoff percentile from the logistic model}
}
\description{
Function to calculate percentiles from the hurdle model
}
