% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculator.R
\name{ScoreCalc}
\alias{ScoreCalc}
\title{Compute IDEFICS Scores for Multiple Variables}
\usage{
ScoreCalc(
  df,
  return_input = F,
  return_values = c("percentile", "z.score", "MetS", "action")
)
}
\arguments{
\item{df}{A data frame with columns: `sex`, `age`, `height`, and observed values for any of the supported variables:
`bmi`, `glu`, `hdl`, `height`, `homa`, `insu`, `trg`, `waist`, `sbp`, `dbp`. Values for `height` are only required for `sbp` and `dbp`.}

\item{return_input}{Logical. If `TRUE`, includes original input columns in the result. Defaults to `FALSE`.}

\item{return_values}{Character vector. Specifies which scores to compute. Options include `"percentile"`, `"z.score"`, `"MetS"`, and `"action"`.}
}
\value{
A data frame with score columns named as `<variable>_percentile`, `<variable>_z.score`, etc.
        Includes additional columns like `MetS` or action levels if requested.
}
\description{
Applies `get_scores()` to multiple anthropometric and metabolic variables in a data frame and optionally returns MetS and action levels.
}
\details{
The function iteratively applies `get_scores()` to each recognized variable in the input and combines the results.
If `"MetS"` is requested, a Metabolic Syndrome score is computed and optionally transformed to percentiles/z-scores.
}
\examples{
ScoreCalc(df, return_values = c("percentile", "action"))

}
