% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculator.R
\name{get_scores}
\alias{get_scores}
\title{Calculate percentiles and z-scores}
\usage{
get_scores(
  variable = "waist",
  sex = c("f", "m"),
  age = 6:5,
  height = NULL,
  values = c(20, 21),
  return_values = c("percentile", "z.score")
)
}
\arguments{
\item{variable}{Either 'waist', 'sbp', 'dbp', 'trg', 'hdl' and 'homa',
'bmi', 'glu', 'height' or 'insu'.}

\item{sex}{A vector of characters 'f' and 'm'.}

\item{age}{A numeric vector.}

\item{height}{A numeric vector.}

\item{values}{A numeric vector with values for the parameter given in `variable`.}

\item{return_values}{A character vector with the possible values 'percentile' and 'z.score'.}
}
\value{
A A list with the values specified in `return_values`
}
\description{
The function calculates the percentiles and z-scores for given value of
several clinical parameters ('waist', 'sbp', 'dbp', 'trg', 'hdl' and 'homa',
'bmi', 'glu', 'height', 'insu')
using the respective sex-specifc (and for some parameters height-specific)
reference table.
}
\examples{
get_scores(variable="dbp", sex=c("f","m"), age=c(5,5), height=c(120,110), values=c(70,60))
}
